fix: set code with comment before merge
