+++ front-end/src/app/[locale]/admin/booking/component/headCell.tsx
@@ -38,10 +38,6 @@ const StatusCell = ({row,content}: {row: Booking,content: string})=>{
       return <span className={`tw-flex tw-w-fit tw-items-center tw-rounded-full tw-py-1 tw-px-4 tw-text-sm tw-font-medium tw-capitalize tw-text-blue-600 tw-bg-blue-100`}>
         {content}
       </span>
-      // case "REFUNDED":
-      //   return <span className={`tw-flex tw-w-fit tw-items-center tw-rounded-full tw-py-1 tw-px-4 tw-text-sm tw-font-medium tw-capitalize tw-text-red-600 tw-bg-red-100`}>
-      //   {content}
-      // </span>
       default:
         break;
     }
