+++ front-end/src/app/[locale]/profile/components/profileContent.tsx
@@ -1,0 +1,44 @@
+"use client";
+
+import React from "react";
+import { Box } from "@mui/material";
+import Divider from "@/components/common/Divider";
+import PersonalProfile from "./PersonalInformation/personalInformation";
+import SecurityInformation from "./SecurityInformation/securityInformation";
+import BookingHistory from "./BookingHistory";
+
+interface ProfileContentProps {
+  activeMenuProfile: number;
+  data: any;
+  avatar?: string;
+  onSuccess: () => void;
+  setAvatar: (avatar: string) => void;
+}
+
+const ProfileContent: React.FC<ProfileContentProps> = ({
+  activeMenuProfile,
+  data,
+  avatar,
+  onSuccess,
+}) => {
+  switch (activeMenuProfile) {
+    case 1:
+      return (
+        <Box>
+          <PersonalProfile
+            data={data}
+            avatar={avatar}
+            onSuccess={onSuccess}
+          />
+          <Divider />
+          <SecurityInformation />
+        </Box>
+      );
+    case 2:
+      return <BookingHistory />;
+    default:
+      return null;
+  }
+};
+
+export default ProfileContent;
